requirements.txt
python-telegram-bot>=20.7,<21.0
psycopg[binary,pool]>=3.1
